% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphExtended.R
\name{extScatterPlot}
\alias{extScatterPlot}
\alias{ggNMDSPlot}
\alias{ggPCAPlot}
\alias{ggPCoAPlot}
\alias{gtNMDSPlot}
\alias{gtPCAPlot}
\alias{gtPCoAPlot}
\title{The Graphs Extended From \code{\link{ggScatterPlot}}}
\usage{
ggNMDSPlot(comm, attr.df, colour.id = NULL, shape.id = NULL,
  link.id = NULL, text.id = NULL, text.or.point = 3,
  text.size.id = NULL, text.size = 3, distance = "bray", k = 2,
  title = "NMDS", title.add.stress = TRUE, stress.digits = 2,
  colour.levels = c(), shape.levels = c(), verbose = TRUE, ...)

gtNMDSPlot(...)

ggPCoAPlot(dist.comm, attr.df, colour.id = NULL, shape.id = NULL,
  link.id = NULL, text.id = NULL, text.or.point = 3,
  text.size.id = NULL, text.size = 3, eig = TRUE, k = 2,
  add.const = FALSE, title = "Classical MDS", verbose = TRUE, ...)

gtPCoAPlot(...)

ggPCAPlot(df.comm, attr.df, x.i = 1, y.i = 2, colour.id = NULL,
  shape.id = NULL, link.id = NULL, text.id = NULL, text.or.point = 3,
  text.data = NULL, scale.pca = TRUE, title = "PCA", verbose = TRUE,
  ...)

gtPCAPlot(...)
}
\arguments{
\item{comm}{Community data for NMDS plot.
It can also be a \code{\link{dist}} object to show the dissimilarity
between pairwised samples in the community data.
Detail to \code{\link{metaMDS}}.}

\item{attr.df}{A data frame of meta data to define
how the plot is coloured, shaped, or linked.
The row names have to match row names of the input data,
such as a matix or \code{\link{dist}} object.
The column names have to contain the id of colour, shape, link, or text.

This data frame will \code{\link{merge}} with data points generated
by PCA or MDS, so that the rows whose names are not matched are dropped.}

\item{colour.id, shape.id, link.id}{The column name in \code{attr.df} to
define how the data points are coloured, shaped, or linked according to values.}

\item{text.id, text.size.id, text.size, title, stress.digits, text.data, colour.levels, shape.levels, verbose}{Refer to \code{\link{ggScatterPlot}}.}

\item{text.or.point}{If 1, then display the text only; if 2, then only points;
if 3, the default, then both the text and points.}

\item{distance}{Dissimilarity index used in \code{\link{vegdist}}.
Default to "bray".}

\item{k}{The number of dimensions in \code{\link{metaMDS}},
or maximum dimension of the space in \code{\link{cmdscale}}. Default to 2.}

\item{title.add.stress}{If TRUE, the default, then add stress from
\code{\link{metaMDS}} to the title.}

\item{...}{Other arguments passed to \code{\link{ggScatterPlot}}.}

\item{dist.comm}{A distance structure such as that returned by \code{\link{dist}},
or a full symmetric matrix containing the dissimilarities for the classical
MDS \code{\link{cmdscale}}.}

\item{eig}{Indicates whether eigenvalues should be returned from \code{\link{cmdscale}}.
Default to TRUE.}

\item{add.const}{Logical indicating in \code{\link{cmdscale}} if an additive
constant c* should be computed, and added to the non-diagonal dissimilarities
such that the modified dissimilarities are Euclidean. Default to FALSE.}

\item{df.comm}{A numeric or complex matrix (or data frame),
which provides the data for the principal components analysis
\code{\link{prcomp}}.
For example, a community data defined by \code{\link{vegan}}
which is a transposed matrix of community matrix in \pkg{ComMA}.
Use either \code{transposeDF} in \pkg{ComMA}
or \code{\link{t}} to rotate the data frame.}

\item{x.i, y.i}{PCA dimension index.}

\item{scale.pca}{Logical indicating in \code{\link{prcomp}}
whether the variables should be scaled to have unit variance
before the analysis takes place. Default to TURE.}
}
\value{
If the function returns a \code{\link{ggplot}} object,
then its name starts with "gg".
It needs to use \code{\link{pdf.ggplot}} to create pdf.
It also keeps the expandability using '+'.

If the function returns a \code{gtable} object, then its name starts with "gt".
This kind of functions use \code{\link{unclip.ggplot}} to turns off clipping for a
\code{\link{ggplot}} object, but returns a \code{gtable} object.
It needs to use \code{\link{pdf.gtable}} to create pdf.
And also \code{\link{plotgt}}
simplifies the code to plot gtable object in console.
}
\description{
The entended graphs from scatter plot inlcude NMDS, PCoA, and PCA.

The difference between principal components analysis (PCA)
and multidimensional scaling (MDS) is discussed in
\url{http://stats.stackexchange.com/questions/14002/whats-the-difference-between-principal-components-analysis-and-multidimensional}.

Principal coordinates analysis (PCoA) is a classical MDS
introduced by (Gower, 1966).
}
\details{
NMDS plot \code{ggNMDSPlot} and \code{gtNMDSPlot} uses \code{\link{metaMDS}}
in \code{\link{vegan}} to create Nonmetric Multidimensional Scaling (NMDS) plot.

Classical MDS \code{ggPCoAPlot} uses \code{\link{cmdscale}} to
create a classical (metric) multidimensional scaling plot of a data matrix.
Also known as principal coordinates analysis (Gower, 1966).

PCA plot \code{gtPCAPlot} uses \code{\link{prcomp}} to create
a principal components analysis (PCA) plot.
}
\note{
Because functions of computing NMDS \code{\link{metaMDS}},
PCoA \code{\link{cmdscale}}, and PCA \code{\link{prcomp}}
take different input data, please make sure you give a valid input to plot.

Set \code{text.repel=T} to use \code{\link{geom_text_repel}}, if too much text labels.

If the points are overlapped to the edge box, then use \code{\link{expand_limits}},
such as \code{gg???(...) + expand_limits(x = c(?, ?), y=c(?, ?))},
to expand the x y axis limits.
}
\references{
The tutorial is avaiable at
\url{https://cdn.rawgit.com/walterxie/gg1L/master/tutorials/OneLinePlotPointsLines.html}
}
\keyword{plot}
\keyword{scatter}

