% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphUtils.R
\name{get_breaks_positive_values}
\alias{get_breaks_positive_values}
\title{Breaks of multiples of 10 for positive values.}
\usage{
get_breaks_positive_values(max.v, start = c(0.1, 1))
}
\arguments{
\item{max.v}{The max value to create breaks.}

\item{start}{The vector of values before 10. Default to c(0.1, 1).}
}
\value{
The vector of multiples of 10 used for breaks.
Mostly used with \code{\link{scientific_10}} together.
}
\description{
Breaks of multiples of 10 for positive values.
}
\examples{
get_breaks_positive_values(max(reads.phyla[,1:6]), start=c(0))

library(ggplot2)
scale_y_continuous(trans = "log", labels = gg1L::scientific_10,
breaks = gg1L::get_breaks_positive_values(max(reads.phyla[,1:6]), start=c(0)))
}
\keyword{utils}

