% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphUtils.R
\name{getMyPalette}
\alias{getMyPalette}
\title{Customised palette with about 70 colours.}
\usage{
getMyPalette(n, add.grey = 0)
}
\arguments{
\item{n}{The number of colors (>= 1) to be in the palette. If missing, then return the whole palette.}

\item{add.grey}{Add grey colours in the begining of palette. Default to 0.}
}
\value{
The customised palette with about 70 colours.
}
\description{
Customised palette with about 70 colours.
}
\examples{
myPalette <- getMyPalette() 
}
\keyword{graph}

