% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg1L.R
\docType{package}
\name{gg1L-package}
\alias{gg1L}
\alias{gg1L-package}
\title{gg1L Introduction}
\description{
This package wraps ggplot2’s code into one-line functions making plot easy.
It also includes some tips and smart code collected from the Internet
to improve plotting,
and few visualisation methods derived from my previous and current works.

Read introduction:
\command{package?gg1L}

See all exported functions:
\command{help(package = "gg1L")}

More details and tutorials are available at \url{https://github.com/walterxie/gg1L}.
}
\author{
Dong Xie, \email{walter@cs.auckland.ac.nz}
}

