% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphAdd.R
\name{ggAddErrorBars}
\alias{ggAddErrorBars}
\title{Add error bars}
\usage{
ggAddErrorBars(lower, upper, dodge.width = 0.9, err.bar.width = 0.25, ...)
}
\arguments{
\item{lower, upper}{The vector of lower or upper plotted in error bars. May use formula 
\code{lower = mean - standard error of mean}, \code{upper = mean + standard error of mean} 
to calculate. \code{standard error of mean = sd(x) / sqrt(length(x))}.}

\item{dodge.width}{Dodging width, when different to the width of the individual elements. 
Default to 0.9. Refer to \code{\link{position_dodge}}.}

\item{err.bar.width}{The width input for \code{\link{geom_errorbar}}.}

\item{...}{Other arguments passed to \code{\link{geom_errorbar}}.}
}
\description{
Add error bars to a \code{\link{ggplot}} object, mostly points \code{\link{geom_point}}. 
Refer to \code{\link{geom_errorbar}}.
}
\examples{
# lower = mean - standard error of mean, upper = mean + standard error of mean
p <- p + ggAddErrorBars(lower, upper)

# Replicate of http://www.r-bloggers.com/building-barplots-with-error-bars/
data(model.test)
myData <- aggregate(model.test$performance, by = list(model = model.test$model, OS = model.test$OS),
    FUN = function(x) c(mean = mean(x), sd = sd(x), n = length(x)))
myData <- do.call(data.frame, myData)
myData$se <- myData$x.sd / sqrt(myData$x.n)
colnames(myData) <- c("model", "OS", "mean", "sd", "n", "se")
bar.chart <- ggBarChart(myData, x.id="model", y.id="mean", fill.id="OS", y.lab="performance")
lower=myData$mean - myData$se
upper=myData$mean + myData$se
bar.chart <- bar.chart + ggAddErrorBars(lower=lower, upper=upper)
bar.chart
}
\keyword{graph}

