% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphAdd.R
\name{ggAddLine}
\alias{ggAddLine}
\title{Add a line}
\usage{
ggAddLine(linetype = 1, xintercept, yintercept, intercept, slope,
  smooth.method, ...)
}
\arguments{
\item{linetype}{Refer to \code{\link{linetype}}:
0 = blank, 1 = solid, 2 = dashed, 3 = dotted,
4 = dotdash, 5 = longdash, 6 = twodash.}

\item{xintercept, yintercept, intercept, slope, smooth.method}{Refer to \code{\link{geom_vline}}, \code{\link{geom_hline}},
\code{\link{geom_abline}}, \code{\link{geom_smooth}}.
They cannot be used at the same time.}

\item{...}{Other arguments passed to \code{\link{layer}}.}
}
\description{
Add a line \code{\link{geom_line}} to a given \code{\link{ggplot}} object.
}
\examples{
ranks.by.group <- data.frame(plot=c("Plot03","Plot02","Plot01"), `16s`=c(3,2,1), `18s`=c(1,2,3), ITS=c(2,1,3), check.names = FALSE)
p <- ggHeatmap(ranks.by.group, melt.id="plot")
p <- p + ggAddLine(linetype = 2, yintercept = 1)
}
\keyword{add}
\keyword{line}

