% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg1LBase.R
\name{ggOptLegend}
\alias{ggOptLegend}
\title{Legend title}
\usage{
ggOptLegend(p, legend.title.fill = NULL, legend.title.colour = NULL,
  legend.title.shape = NULL, legend.title.group = NULL,
  legend.title.size = NULL, legend.col = 1, legend.row = 0,
  no.legend = NULL, verbose = TRUE)
}
\arguments{
\item{legend.title.fill, legend.title.colour, legend.title.shape, legend.title.group, }{The title of legend is created by fill, colour, shape, or group. Only choose one.
Set legend.title.*="" to remove legend.}

\item{legend.title.size}{The font size of legend title}

\item{legend.col, legend.row}{Customize the number of columns or rows for legend in bar chart.
They cannot be used at the same time. Default not to use them, legend.col=1, legend.row=0.}

\item{no.legend}{Turning off some legends, such as, fill, shape, colour.
Default to NULL, which is to do nothing.}
}
\description{
Change the legend title according to the chart type.
For example, \code{legend.title.fill} is for the colour
related aesthetic \code{\link{fill}} in \pkg{ggplot2},
such as bars.
\code{legend.title.colour} for \code{\link{colour}}, such as line or text.
\code{legend.title.shape} for \code{\link{shape}}, such as shapes of points.
\code{legend.title.group} for \code{\link{group}}.
}

