% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg1LBase.R
\name{ggOptPalette}
\alias{ggOptPalette}
\title{Colour Schemes}
\usage{
ggOptPalette(p, scale.to = "colour", scale.type = NULL, palette = NULL,
  limits = NULL, verbose = TRUE)
}
\arguments{
\item{p}{The ggplot object.}

\item{scale.to}{Choose either "colour" or "fill" to make up
a full command, such as \code{\link{scale_colour_gradient}}.}

\item{scale.type}{Choose one from "brewer", "gradientn", and "manual"
to make up a full command, such as \code{\link{scale_fill_manual}}.}

\item{palette}{The colour palette for bar, box, scatter plot, etc.}

\item{limits}{A numeric vector of length two providing limits of the scale.
Use NA to refer to the existing minimum or maximum.
Refer to \code{\link{continuous_scale}}.}
}
\description{
\code{ggOptPalette} provides colour schemes for graphs.
}
\details{
It uses \code{scale.to}, \code{scale.type}, and \code{palette}
to determin which colour scale is applied.
If \code{scale.type} is NULL as default, then the length of
\code{palette} will be used to the colour scale:

If \code{palette} length == 1, then use \code{\link{scale_*_brewer}}
(\url{http://www.datavis.ca/sasmac/brewerpal.html}), such as "Set1" (max 8 colours).
If length != 1 and length <= 3, then use \code{\link{scale_colour_gradientn}},
such as c("blue", "orange").
Otherwise \code{\link{scale_fill_manual}} will be used given a vector of customized colours.
Set \code{palette=c("red","darkgreen","darkgrey","")}, for example,
if \code{\link{scale_fill_manual}} has to be used when the actual length(palette) <=3.
}
\examples{
p <- ggOptPalette(p, scale.to="fill", palette="Set1")
p <- ggOptPalette(p, scale.to="colour", palette=c("blue", "orange"))
}

