% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg1LBase.R
\name{ggOptScaleAxis}
\alias{ggOptScaleAxis}
\title{Position scales}
\usage{
ggOptScaleAxis(p, axis = "y", scale = "continuous", trans = "identity",
  expand = waiver(), breaks = waiver(), labels = waiver(),
  auto.scale.max = NULL, breaks.start = c(0), verbose = TRUE)
}
\arguments{
\item{p}{The ggplot object.}

\item{axis}{"x" or "y".}

\item{scale}{Choose either "continuous" or "discrete".}

\item{trans}{The string defines the data scale used in either x-axis or y-axis,
which can be "identity" standing for normal, or "per" standing for percentage,
moreover either the name of a transformation object for \code{\link{scale_x_continuous}}
or \code{\link{scale_y_continuous}} (e.g. \code{trans="log"}), or the object itself.
Built-in transformations include "asn", "atanh", "boxcox", "exp", "identity",
"log", "log10", "log1p", "log2", "logit", "probability", "probit", "reciprocal",
"reverse" and "sqrt". Default to "identity".}

\item{expand, breaks, labels}{Refer to \code{\link{scale_x_continuous}} or \code{\link{scale_x_discrete}}.}

\item{auto.scale.max}{If not NULL, then use \code{\link{get_breaks_positive_values}}
to create breaks regarding to the given maximun value.}

\item{breaks.start}{Refer to \code{\link{get_breaks_positive_values}}.
Use \code{start=c(0.1, 1)} to show 1 in a log-scaled axis.}
}
\description{
\code{ggOptScaleAxis} provides scales for discrete or continuous x and y aesthetics.
}
\examples{
p <- ggOptPalette(p, axis="y", scale="continuous", trans="per")
p <- ggOptPalette(p, scale.to="colour", palette=c("blue", "orange"))
}

