% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphExtendedBarChart.R
\name{ggYAcrossXBar}
\alias{ggYAcrossXBar}
\title{Y Across X Bar Chart}
\usage{
ggYAcrossXBar(community.matrix, terms = c("OTUs", "samples", "reads"),
  title = "The number of OTUs/reads across the number of samples",
  x.lab = "Number of samples crossed", y.lab = "Number of OTUs/reads",
  y.trans = "log", auto.scale.y = TRUE, x.scale = "discrete",
  x.interval = 1, x.text.angle = 0, legend.title = "", ...)
}
\arguments{
\item{community.matrix}{Community matrix (OTU table), where rows are 
OTUs or individual species and columns are sites or samples.}

\item{terms}{The terms of matrix for x, y, and cells. 
Default to \code{c("OTUs", "samples", "reads")}.}

\item{title}{Graph title}

\item{x.lab, y.lab}{The label of x-axis or y-axis, such as plot names.}

\item{...}{Other arguments passed to \code{\link{ggBarChart}}.}

\item{low, }{high Refer to \pkg{ggplot2} \code{\link{scale_fill_gradient}}. 
Default to low="white", high="steelblue".}

\item{autoWidth}{If TRUE, then use number of bars and legend columns 
to estimate pdf width automatically. Default to TRUE.}
}
\description{
The bar chart shows the number of OTUs/reads in the bar, 
which simultaneously appeared at the number of samples in the value of x-axis. 
Given a data matrix whose structure is same as community matrix defined in \code{\link{ComMA}}, 
this function uses \code{\link{cmYAcrossX}} to aggregate it into another abundance matrix, 
and uses \code{\link{ggBarChart}} to plot.
}
\examples{
 
community.matrix <- getCommunityMatrix("16S", isPlot=TRUE, minAbund=1)
bar.yx <- ggYAcrossXBar(community.matrix)
}
\keyword{graph}

