% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphUtils.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{\code{grid_arrange_shared_legend} shares a legend
between multiple plots using \code{\link{grid.arrange}}.
Return a \code{\link{gtable}} object.}
\usage{
grid_arrange_shared_legend(..., input.list = FALSE,
  legend.position = "bottom", ncol = 2, nrow = 2, widths = c(1, 0.1),
  unclip.ggplot = TRUE)
}
\arguments{
\item{...}{The list of \code{\link{ggplot}} objects.}

\item{input.list}{Default to FALSE to unwrap list(...) to
get the actual list if the input is a list of plots}

\item{legend.position}{The position of legends
("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{ncol, nrow}{Specify the grid.}

\item{widths}{A unit vector giving the width of each column in \code{\link{gtable}}.
\code{length(widths) == ncol}, for example, widths=c(1, 0.1, 0.1) for 3 columns.}
}
\description{
Modified from
\url{http://rpubs.com/sjackman/grid_arrange_shared_legend}.
}
\examples{
library(ggplot2)
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
p1 <- qplot(carat, price, data=dsamp, colour=clarity)
p2 <- qplot(cut, price, data=dsamp, colour=clarity)
p3 <- qplot(color, price, data=dsamp, colour=clarity)
p4 <- qplot(depth, price, data=dsamp, colour=clarity)
grid_arrange_shared_legend(p1, p2, p3, p4)

grid_arrange_shared_legend(list(p1, p2, p3, p4), input.list=TRUE)
}
\keyword{graph}

