% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphExtended.R
\name{gtRarefactionCurve}
\alias{gtRarefactionCurve}
\title{Rarefaction curves for multi-sample, which is extended from \code{\link{gtLine}}.}
\usage{
gtRarefactionCurve(df.size, attr.df, group.id = "Samples", colour.id = NULL,
  line.type = 2, line.alpha = 0.75, shape.id = NULL,
  x.prefix = "^.*?\\\\.", line.or.point = 3, end.point.only = TRUE,
  title = "Rarefaction Curves", x.lab = "Reads", y.lab = "Diversity",
  verbose = TRUE, ...)
}
\arguments{
\item{df.size}{A data frame required to \code{\link{merge}} with \code{attr.df} 
and \code{\link{melt}} before making a group of rarefaction curves.
The rows are samples and must be a subset of rows in \code{attr.df}, 
columns are the subsampled data points to draw the curve. 
The data frame of rarefaction curve of phylogenetic diversity can be generated by 
\code{\link{getPhylorareDF}}. 
For example,
\tabular{rrrrr}{
  Samples \tab size.1 \tab size.5 \tab size.67 \tab ...\cr
  Sample1 \tab 1.845445 \tab 3.679956 \tab 9.191672 \tab ...\cr
  Sample2 \tab 2.047155 \tab 10.41827 \tab 17.34067 \tab ...\cr
  Sample3 \tab 0.06646017 \tab 1.65030905 \tab NaN \tab ... 
}}

\item{attr.df}{A data frame to define how rarefaction curves are coloured, shaped.}

\item{group.id}{A column name from \code{df.to.melt} to \code{\link{melt}} 
subsampled data points, which is copied from merged "row.names".}

\item{x.prefix}{The regular expression to remove prefix from column names in 
\code{df.to.melt} (e.g. size.100). Default to "^.*?\\\\.".}

\item{line.or.point}{If 1, then display the line only; if 2, then only points; 
if 3, the default, then both the line and points.}

\item{end.point.only}{If TRUE, as default, then only display the points in 
the end of lines. Otherwise, display all the points.}

\item{...}{Other arguments passed to \code{\link{gtLine}}.}
}
\description{
Rarefaction curves for multi-sample, which is extended from \code{\link{gtLine}}.
}
\examples{
rare.curv <- gtRarefactionCurve(df.size, attr.df, group.id="Samples", colour.id="Species",  
                       shape.id="GutSegment",point.size=2, x.trans="log", auto.scale.x=T)
plot(rare.curv)
}
\keyword{graph}

