% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmc.log}
\alias{mcmc.log}
\title{*BEAST log.}
\format{A data frame with 2001 rows and 18 columns:
\describe{
  \item{row.names}{the state number in MCMC}
  \item{posterior}{the posterior}
  ...
  \item{TreeHeight.Species}{the species tree height}
  ...
}}
\source{
\url{http://beast2.org/}
}
\usage{
mcmc.log
}
\description{
A MCMC log from the *BEAST \url{mbe.oxfordjournals.org/content/27/3/570.full} model,
whose row names are the states, column names are the parameters.
}
\keyword{datasets}

