% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model.test}
\alias{model.test}
\title{Performance testing.}
\format{A data frame with 180 rows and 4 columns:
\describe{
  \item{model}{a GTR model}
  \item{test}{test ID}
  \item{performance}{the performance score}
  \item{OS}{operation system}
}}
\source{
\url{http://www.beast2.org/2016/04/05/beast-1-vs-2-performance-benchmarking/}
}
\usage{
model.test
}
\description{
A dataset containing the performance score of tests
in four GTR models in three operation systems using BEAST.
}
\keyword{datasets}

