% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphUtils.R
\name{mylog_trans}
\alias{mylog_trans}
\title{Defining the scale change.
It is mostly used to force bars to start from a lower value than 0 in \pkg{ggplot2} \code{\link{geom_bar}} in R}
\source{
\url{http://stackoverflow.com/questions/22295253/force-bars-to-start-from-a-lower-value-than-0-in-ggplot-geom-bar-in-r}
}
\usage{
mylog_trans(base = exp(1), from = 0)
}
\arguments{
\item{base}{The base of logarithm to use. Default to exp(1).}

\item{from}{The value to start from. Default to 0.}
}
\value{
The scale.
}
\description{
Defining the scale change.
It is mostly used to force bars to start from a lower value than 0 in \pkg{ggplot2} \code{\link{geom_bar}} in R
}
\examples{
# starts from 1e-2
scale_y_continuous(trans = mylog_trans(base=10, from=-2))
}
\keyword{graph}

