% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphUtils.R
\name{pdf}
\alias{pdf}
\alias{pdf.ggplot}
\alias{pdf.gtable}
\alias{pdf.plot}
\alias{plot.gtable}
\alias{unclip.ggplot}
\title{Create pdf}
\usage{
pdf.ggplot(gg.plot, fig.path, width = 6, height = 6, useDingbats = TRUE)

pdf.gtable(g.table, fig.path, width = 6, height = 6, useDingbats = TRUE)

pdf.plot(plot, fig.path, width = 6, height = 6, useDingbats = TRUE)

\method{plot}{gtable}(g.table)

unclip.ggplot(gg.plot)
}
\arguments{
\item{gg.plot}{A \code{\link{ggplot}} object.}

\item{fig.path}{fig.path The full path of pdf file.}

\item{width, }{height Refer to \code{\link{pdf}}. Default to width=6, height=6.}

\item{useDingbats}{Defaults to TRUE, which produces smaller and better output.
Setting this to FALSE can work around font display problems in broken PDF viewers.
Refer to \code{\link{pdf}}.}

\item{plot}{The list returned from \code{\link{plot}}.}

\item{gtable}{A \code{\link{gtable}} object.}
}
\details{
\code{pdf.ggplot} creates pdf for ggplot object.

\code{pdf.gtable} creates pdf for gtable object.

\code{pdf.plot} creates pdf for the list returned from \code{\link{plot}}.
Tip: use \code{\link{"\%<a-\%"}} in \pkg{pryr} to save plots.

\code{plot.gtable} \code{\link{require}(\link{grid})}
and plot gtable object in console.

\code{unclip.ggplot} turns off clipping for a
\code{\link{ggplot}} object, but returns a
\code{\link{gtable}} object.
Refer to \url{http://stackoverflow.com/questions/9690648/point-clipped-on-x-axis-in-ggplot}.
}
\examples{
pdf.ggplot(gg.plot, fig.path="fig.pdf", width=8, height=8)

pdf.gtable(g.table, fig.path="fig.pdf", width=8, height=8)

pdf.plot(plot, fig.path="fig.pdf", width=8, height=8)

plot(g.table)

gt <- unclip.ggplot(gg.plot)

}
\keyword{graph}

