% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphUtils.R
\name{scientific_10}
\alias{scientific_10}
\title{Scientific notation}
\usage{
scientific_10(x)
}
\arguments{
\item{x}{The number (not string type).}
}
\value{
The mathematical \code{\link{expression}} in scientific notation of a given number.
}
\description{
Scientific notation
}
\examples{
#expression(10^04)
scientific_10(10000)
}
\keyword{graph}

