% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsUtils.R
\name{StatsUtils}
\alias{StatsUtils}
\alias{cmYAcrossX}
\alias{random2Clusters}
\alias{summarySE}
\title{Data and Statistics Utils}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)

cmYAcrossX(community.matrix, terms = c("OTUs", "samples", "reads"))

random2Clusters(n = 100, seed = 101)
}
\arguments{
\item{data}{A data frame.}

\item{measurevar}{the name of a column that contains the variable
to be summariezed.}

\item{groupvars}{a vector containing names of columns that
contain grouping variables.}

\item{na.rm}{a boolean that indicates whether to ignore NA's.}

\item{conf.interval}{the percent range of the confidence interval,
such as .95 as default.}

\item{.drop}{should combinations of variables that do not appear
in the input data be preserved (FALSE) or dropped (TRUE, default).}

\item{community.matrix}{\strong{Community matrix} in \pkg{ComMA} is a matrix or data.frame,
where rows are OTUs (therefore this matrix is called \emph{OTU table}) or individual species
and columns are sites or samples.
Matrix elements are abundance data or proportion (e.g. counts, percentage). For example,
\tabular{rrrr}{
  OTU_id \tab plot01 \tab plot02\tab ...\cr
  OTU_1 \tab 1 \tab 0 \tab ...\cr
  OTU_2 \tab 100 \tab 200 \tab ...\cr
  OTU_3 \tab 56 \tab 3 \tab ...
}}

\item{terms}{The terms to be used in names of the data frame,
which will be shown in the graph if using \code{\link{ggYAcrossXBar}}.
Default to c("OTUs", "samples", "reads").
Please be careful of the order if any change.}

\item{n}{The number of points. Default to 100.}

\item{seed}{An integer seed for \code{\link{set.seed}}. Default to 101.}
}
\description{
Useful functions to summarizse data.
}
\details{
\code{summarySE} gives count, mean, standard deviation,
standard error of the mean, and confidence interval (default 95%).
The code is copied from
\url{http://www.cookbook-r.com/Manipulating_data/Summarizing_data/}.

\code{cmYAcrossX} aggregates a community matrix to another abundance matrix
to show the number of OTUs (y-axis) simultaneously appeared at the number of samples (x-axis).
The 'samples' is the number of samlpes listed in sequence,
the 'OTUs' is the number of OTUs simultaneously appeared only in that number of samlpes,
and the 'reads' is the number of reads assigned to those OTUs.

\code{random2Clusters} generates coordinates for 2 clusters.
\url{http://stackoverflow.com/questions/2397097/how-can-a-data-ellipse-be-superimposed-on-a-ggplot2-scatterplot}.
}
\examples{
data(model.test)
summarySE(model.test, measurevar="performance", groupvars=c("model", "OS"))
# With a data set with NA's, use na.rm=TRUE
dataNA <- model.test
dataNA$performance[11:15] <- NA
summarySE(dataNA, measurevar="performance", groupvars=c("model", "OS"), na.rm=TRUE)

data(reads.phyla)
cm.aggre <- cmYAcrossX(reads.phyla[,1:6])
cm.aggre

df.clusters <- random2Clusters()

}
\keyword{statistics}
\keyword{utils}

