% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphExtended.R
\name{extScatterPlot}
\alias{extScatterPlot}
\alias{ggNMDSPlot}
\alias{gtClassicalMDSPlot}
\alias{gtNMDSPlot}
\alias{gtPCAPlot}
\title{The Graphs Extended From \code{\link{ggScatterPlot}}}
\usage{
ggNMDSPlot(comm, attr.df, colour.id = NULL, shape.id = NULL,
  link.id = NULL, text.id = NULL, text.or.point = 3,
  text.size.id = NULL, text.size = 3, distance = "bray", k = 2,
  title = "NMDS", title.add.stress = TRUE, stress.digits = 2,
  colour.levels = c(), shape.levels = c(), verbose = TRUE, ...)

gtNMDSPlot(comm, attr.df, ...)

gtClassicalMDSPlot(dist.comm, attr.df, colour.id = NULL, shape.id = NULL,
  link.id = NULL, text.id = NULL, text.or.point = 3,
  text.size.id = NULL, text.size = 3, eig = TRUE, k = 2,
  add.const = FALSE, title = "Classical MDS", verbose = TRUE, ...)

gtPCAPlot(df.comm, attr.df, x.i = 1, y.i = 2, colour.id = NULL,
  shape.id = NULL, link.id = NULL, text.id = NULL, text.or.point = 3,
  text.data = NULL, scale.pca = TRUE, title = "PCA", verbose = TRUE,
  ...)
}
\arguments{
\item{comm}{Community data for NMDS plot. 
It is dissimilarities either as a \code{\link{dist}} structure 
or as a community data defined by \code{\link{vegan}} which is 
a transposed matrix of community matrix in \code{\link{ComMA}}.
Use \code{\link{transposeDF}} to rotate the data frame.
Detail to \code{\link{metaMDS}}.}

\item{attr.df}{A data frame of meta data to define 
how the plot is coloured, shaped, or linked.
The row names have to match row names of the input data, 
such as a matix or \code{\link{dist}} object.
The column names have to contain the id of colour, shape, link, or text.

This data frame will \code{\link{merge}} with data points generated 
by PCA or MDS, so that the rows whose names are not matched are dropped.}

\item{colour.id, shape.id, link.id}{The column name in \code{attr.df} to 
define how the data points are coloured, shaped, or linked according to values.}

\item{text.or.point}{If 1, then display the text only; if 2, then only points; 
if 3, the default, then both the text and points.}

\item{distance}{Dissimilarity index used in \code{\link{vegdist}}.
Default to "bray".}

\item{k}{The number of dimensions in \code{\link{metaMDS}}, 
or maximum dimension of the space in \code{\link{cmdscale}}. Default to 2.}

\item{title.add.stress}{If TRUE, the default, then add stress from 
\code{\link{metaMDS}} to the title.}

\item{...}{Other arguments passed to \code{\link{ggScatterPlot}}.}

\item{dist.comm}{A distance structure such as that returned by \code{\link{dist}},
or a full symmetric matrix containing the dissimilarities for the classical 
MDS \code{\link{cmdscale}}.}

\item{eig}{Indicates whether eigenvalues should be returned from \code{\link{cmdscale}}. 
Default to TRUE.}

\item{add.const}{Logical indicating in \code{\link{cmdscale}} if an additive 
constant c* should be computed, and added to the non-diagonal dissimilarities 
such that the modified dissimilarities are Euclidean. Default to FALSE.}

\item{df.comm}{A numeric or complex matrix (or data frame), 
which provides the data for the principal components analysis 
\code{\link{prcomp}}.
For example, a community data defined by \code{\link{vegan}} which is 
a transposed matrix of community matrix in \code{\link{ComMA}}.
Use \code{\link{transposeDF}} to rotate the data frame.}

\item{scale.pca}{Logical indicating in \code{\link{prcomp}} 
whether the variables should be scaled to have unit variance 
before the analysis takes place. Default to TURE.}
}
\value{
A \code{\link{gtable}} object, which turns off clipping.
It needs to use \code{\link{pdf.gtplot}} to create pdf, 
or \code{\link{plot.gtable}} to plot in console.
}
\description{
The difference between principal components analysis (PCA) 
and multidimensional scaling (MDS) is discussed in 
\url{http://stats.stackexchange.com/questions/14002/
whats-the-difference-between-principal-components-analysis-and-multidimensional}.

A classical MDS is also known as principal coordinates analysis (Gower, 1966).
}
\details{
NMDS plot \code{ggNMDSPlot} and \code{gtNMDSPlot} uses \code{\link{metaMDS}} 
in \code{\link{vegan}} to create Nonmetric Multidimensional Scaling (NMDS) plot.

Classical MDS \code{gtClassicalMDSPlot} uses \code{\link{cmdscale}} to  
create a classical (metric) multidimensional scaling plot of a data matrix. 
Also known as principal coordinates analysis (Gower, 1966).

PCA plot \code{gtPCAPlot} uses \code{\link{prcomp}} to create 
a principal components analysis (PCA) plot.
}
\note{
Because functions of computing NMDS \code{\link{metaMDS}}, 
PCoA \code{\link{cmdscale}}, and PCA \code{\link{prcomp}} 
take different input data, please make sure you give a valid input to plot.

Set \code{text.repel=T} to use \code{\link{geom_text_repel}}, if too much text labels.

If the points are overlapped to the edge box, then use \code{\link{expand_limits}}, 
such as \code{gg???(...) + expand_limits(x = c(?, ?), y=c(?, ?))},
to expand the x y axis limits.
}
\examples{
# display text only and use ggrepel, also expand the x y axis limits
nmds.plot <- ggNMDSPlot(correlations, text.or.point=1, text.repel=T) + expand_limits(x = c(-1, 1), y=c(-1, 1)) 
# turns off clipping
nmds.plot <- gtNMDSPlot(comm, env, colour.id="FishSpecies", shape.id="FeedingPattern", text.or.point=2)
plot(nmds.plot)

cmds.plot <- gtClassicalMDSPlot(dist.comm, env, colour.id="FishSpecies", shape.id="FeedingPattern")
plot(cmds.plot)

pca.plot <- gtPCAPlot(df.comm, env, colour.id="FishSpecies", shape.id="FeedingPattern")
plot(pca.plot)

}
\keyword{graph}

