% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphUtils.R
\name{get_breaks_positive_values}
\alias{get_breaks_positive_values}
\title{Breaks of multiples of 10 for positive values.}
\usage{
get_breaks_positive_values(max.v, start = c(0.1, 1))
}
\arguments{
\item{max.v}{The max value to create breaks.}

\item{start}{The vector of values before 10. Default to c(0.1, 1).}
}
\value{
The vector of multiples of 10 used for breaks. 
Mostly used with \code{\link{scientific_10}} together.
}
\description{
Breaks of multiples of 10 for positive values.
}
\examples{
get_breaks_positive_values(68759)
[1] 1e-01 1e+00 1e+01 1e+02 1e+03 1e+04 1e+05

scale_y_continuous(trans = "log", labels = ComMA::scientific_10, 
breaks = ComMA::get_breaks_positive_values(max(df, start=c(0))))
}
\keyword{graph}

