% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reads.phyla}
\alias{reads.phyla}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 110 rows and 7 columns,
the last column is the taxonomy group:
\describe{
  \item{16S}{a genetic mark}
  ...
  \item{TaxaGroup}{the taxonomy group of phyla}
}}
\source{
\url{http://dx.doi.org/10.1186/s13742-015-0086-1/}
}
\usage{
reads.phyla
}
\description{
A dataset containing the number of eDNA sequences (reads)
identified as the phyla listed in row names,
using a genetic mark in each column.
}
\keyword{datasets}

