% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsUtils.R
\name{StatsUtils}
\alias{StatsUtils}
\alias{summarySE}
\title{Data and Statistics Utils}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{measurevar}{the name of a column that contains the variable 
to be summariezed.}

\item{groupvars}{a vector containing names of columns that 
contain grouping variables.}

\item{na.rm}{a boolean that indicates whether to ignore NA's.}

\item{conf.interval}{the percent range of the confidence interval,
such as .95 as default.}
}
\description{
Useful functions to summarizse data.
}
\details{
\code{summarySE} gives count, mean, standard deviation, 
standard error of the mean, and confidence interval (default 95%).
The code is copied from 
\url{http://www.cookbook-r.com/Manipulating_data/Summarizing_data/}.
}
\examples{
summarySE(model.test, measurevar="performance", groupvars=c("model", "OS"))
# With a data set with NA's, use na.rm=TRUE
summarySE(data, measurevar="change", groupvars=c("sex", "condition"), na.rm=TRUE)

}
\keyword{statistics}

